Move `equals` json helper outside of diff
Change-type: patch
