+++ lib/json.ts
@@ -1,0 +1,27 @@
+function isObject(value: unknown): value is object {
+	return typeof value === 'object' && value !== null;
+}
+
+/**
+ * Calculates deep equality between javascript
+ * objects
+ */
+export function equals<T>(value: T, other: T): boolean {
+	if (isObject(value) && isObject(other)) {
+		const [vProps, oProps] = [value, other].map(
+			(a) => Object.getOwnPropertyNames(a) as Array<keyof T>,
+		);
+		if (vProps.length !== oProps.length) {
+			// If the property lists are different lengths we don't need
+			// to check any further
+			return false;
+		}
+
+		// Otherwise this comparison will catch it. This works even
+		// for arrays as getOwnPropertyNames returns the list of indexes
+		// for each array
+		return vProps.every((key) => equals(value[key], other[key]));
+	}
+
+	return value === other;
+}
